Removed models.
